refactor(nns): move tests
